+++ src/coreclr/tools/aot/Mono.Linker.Tests/TestCasesRunner/TestCaseLinkerOptions.cs
@@ -18,7 +18,6 @@ namespace Mono.Linker.Tests.TestCasesRunner
 		public bool IgnoreSubstitutions;
 		public bool IgnoreLinkAttributes;
 		public string KeepTypeForwarderOnlyAssemblies;
-		public string KeepDebugMembers;
 		public string LinkSymbols;
 		public bool SkipUnresolved;
 		public bool StripDescriptors;
