+++ src/tools/illink/test/Mono.Linker.Tests.Cases/Attributes.Debugger/DebuggerDisplayAttributeOnType.cs
@@ -4,17 +4,7 @@ using Mono.Linker.Tests.Cases.Expectations.Metadata;
 
 namespace Mono.Linker.Tests.Cases.Attributes.Debugger
 {
-#if NETCOREAPP
 	[SetupLinkAttributesFile ("DebuggerAttributesRemoved.xml")]
-#else
-	[SetupLinkerTrimMode ("link")]
-	[SetupLinkerKeepDebugMembers ("false")]
-
-	// Can be removed once this bug is fixed https://bugzilla.xamarin.com/show_bug.cgi?id=58168
-	[SkipPeVerify (SkipPeVerifyForToolchian.Pedump)]
-
-	[KeptMemberInAssembly (PlatformAssemblies.CoreLib, typeof (DebuggerDisplayAttribute), ".ctor(System.String)")]
-#endif
 	public class DebuggerDisplayAttributeOnType
 	{
 		public static void Main ()
@@ -26,9 +16,6 @@ namespace Mono.Linker.Tests.Cases.Attributes.Debugger
 
 		[Kept]
 		[KeptMember (".ctor()")]
-#if !NETCOREAPP
-		[KeptAttributeAttribute (typeof (DebuggerDisplayAttribute))]
-#endif
 		[DebuggerDisplay ("{Property}")]
 		class Foo
 		{
@@ -37,9 +24,6 @@ namespace Mono.Linker.Tests.Cases.Attributes.Debugger
 
 		[Kept]
 		[KeptMember (".ctor()")]
-#if !NETCOREAPP
-		[KeptAttributeAttribute (typeof (DebuggerDisplayAttribute))]
-#endif
 		[DebuggerDisplay ("{Method()}")]
 		class Bar
 		{
@@ -51,9 +35,6 @@ namespace Mono.Linker.Tests.Cases.Attributes.Debugger
 
 		[Kept]
 		[KeptMember (".ctor()")]
-#if !NETCOREAPP
-		[KeptAttributeAttribute (typeof (DebuggerDisplayAttribute))]
-#endif
 		[DebuggerDisplay (null)]
 		class Baz
 		{
