+++ src/tools/illink/test/Mono.Linker.Tests.Cases/Attributes.Debugger/DebuggerDisplayAttributeOnTypeThatIsNotUsed.cs
@@ -1,18 +1,9 @@
 ﻿using System.Diagnostics;
-using Mono.Linker.Tests.Cases.Expectations.Assertions;
 using Mono.Linker.Tests.Cases.Expectations.Metadata;
 
 namespace Mono.Linker.Tests.Cases.Attributes.Debugger
 {
-#if NETCOREAPP
 	[SetupLinkAttributesFile ("DebuggerAttributesRemoved.xml")]
-#else
-	[SetupLinkerTrimMode ("link")]
-	[SetupLinkerKeepDebugMembers ("false")]
-
-	// Can be removed once this bug is fixed https://bugzilla.xamarin.com/show_bug.cgi?id=58168
-	[SkipPeVerify (SkipPeVerifyForToolchian.Pedump)]
-#endif
 	public class DebuggerDisplayAttributeOnTypeThatIsNotUsed
 	{
 		public static void Main ()
