--- src/tools/illink/test/Mono.Linker.Tests.Cases/Attributes.Debugger/DebuggerDisplayAttributeOnTypeWithNonExistentMethod.cs
@@ -1,39 +1,0 @@
-﻿#define FLAG
-
-using System.Diagnostics;
-using Mono.Linker.Tests.Cases.Expectations.Assertions;
-using Mono.Linker.Tests.Cases.Expectations.Metadata;
-
-namespace Mono.Linker.Tests.Cases.Attributes.Debugger
-{
-#if !NETCOREAPP
-	[SetupLinkerTrimMode ("link")]
-	[SetupLinkerKeepDebugMembers ("false")]
-
-	// Can be removed once this bug is fixed https://bugzilla.xamarin.com/show_bug.cgi?id=58168
-	[SkipPeVerify (SkipPeVerifyForToolchian.Pedump)]
-
-	[KeptMemberInAssembly (PlatformAssemblies.CoreLib, typeof (DebuggerDisplayAttribute), ".ctor(System.String)")]
-#endif
-	public class DebuggerDisplayAttributeOnTypeWithNonExistentMethod
-	{
-		public static void Main ()
-		{
-			var bar = new Bar ();
-		}
-
-		[Kept]
-		[KeptMember (".ctor()")]
-		[KeptAttributeAttribute (typeof (DebuggerDisplayAttribute))]
-		[DebuggerDisplay ("{Method()}")]
-		class Bar
-		{
-#if !FLAG
-			public int Method ()
-			{
-				return 1;
-			}
-#endif
-		}
-	}
-}
