+++ src/tools/illink/test/Mono.Linker.Tests/TestCasesRunner/LinkerArgumentBuilder.cs
@@ -106,12 +106,6 @@ namespace Mono.Linker.Tests.TestCasesRunner
 			Append (value);
 		}
 
-		public virtual void AddKeepDebugMembers (string value)
-		{
-			Append ("-v");
-			Append (value);
-		}
-
 		public virtual void AddAssemblyAction (string action, string assembly)
 		{
 			Append ("--action");
@@ -218,9 +212,6 @@ namespace Mono.Linker.Tests.TestCasesRunner
 
 			if (!string.IsNullOrEmpty (options.LinkSymbols))
 				AddLinkSymbols (options.LinkSymbols);
-
-			if (!string.IsNullOrEmpty (options.KeepDebugMembers))
-				AddKeepDebugMembers (options.KeepDebugMembers);
 
 			AddSkipUnresolved (options.SkipUnresolved);
